util: invert a dict with list values
Give a dict with list values, make each elem of each list a key that points back ot the key of its list. Assumes lists are mutex

I use this to invert the character_bins.json file. I do it in this twisted way because it makes maintaining character_bins.json a lot easier, but greatly improves the speed of using that file.